Find Top Ten Users
Given the list of user information, print out the top ten users based on their reputation. In your code,
each mapper determines the top ten records of its input split and outputs them to the reduce phase. The
mappers are filtering their input split to the top ten records, and the reducer is responsible for the final
ten.